+++ src/components/atoms/HookFormInputWrapper/HookFormInputWrapper.stories.tsx
@@ -1,11 +1,13 @@
 import { StoryObj } from "@storybook/react";
-import { useWatch } from "react-hook-form";
+import { useState } from "react";
+import { useForm } from "react-hook-form";
 import { JSONTree } from "react-json-tree";
 import { SetRequired } from "type-fest";
 
 import HookForm from "../HookForm.js";
 import NumberInput, { NumberInputProps } from "../NumberInput/NumberInput.js";
 import PhoneNumberInput from "../PhoneNumberInput/PhoneNumberInput.js";
+import { Button } from "../index.js";
 import HookFormInputWrapper from "./HookFormInputWrapper.js";
 
 const metadata = {
@@ -27,8 +29,15 @@ function InputWrapper({
   );
 }
 
-function FormContent() {
-  const watchAll = useWatch();
+const FormContent = () => {
+  const { setValue, watch } = useForm({
+    defaultValues: {
+      "NumberInput (format with default)": "1234.1003",
+    },
+    mode: "onChange",
+  });
+  const [updateValue, setUpdateValue] = useState(2000.001);
+  const watchAll = watch();
   return (
     <div>
       <div
@@ -41,6 +50,10 @@ function FormContent() {
       >
         <InputWrapper name="NumberInput" />
         <InputWrapper format name="NumberInput (format)" />
+        <InputWrapper
+          format={{ maximumFractionDigits: 4 }}
+          name={"NumberInput (format with default)"}
+        />
         <InputWrapper
           emptyValue={null}
           format
@@ -60,6 +73,16 @@ function FormContent() {
             <input />
           </HookFormInputWrapper>
         </div>
+        <div>
+          <Button
+            onClick={() => {
+              setValue("NumberInput (format with default)", updateValue + 0.01);
+              setUpdateValue(updateValue + 0.01);
+            }}
+          >
+            Update Value
+          </Button>
+        </div>
       </div>
       <div style={{ fontSize: ".75rem", fontWeight: 500, marginTop: "1rem" }}>
         <div>Values:</div>
@@ -67,7 +90,7 @@ function FormContent() {
       </div>
     </div>
   );
-}
+};
 
 const Template = () => {
   return (