+++ src/tools/number.ts
@@ -13,7 +13,7 @@ export function formatNumber(
     minimumFractionDigits,
     nullValue,
     ...options
-  }: FormatNumberOptions = {},
+  }: FormatNumberOptions = {}
 ) {
   const number = stringToNumber(value, nullValue);
   if (!isFinite(number)) {
@@ -21,14 +21,13 @@ export function formatNumber(
   }
   const adjustedMinFractionDigits = Math.max(
     minimumFractionDigits ?? fractionDigits,
-    isString(value) && value.endsWith(".") ? 1 : 0,
+    isString(value) && value.endsWith(".") ? 1 : 0
   );
   return number?.toLocaleString("en-IN", {
     maximumFractionDigits: Math.max(
       maximumFractionDigits ?? fractionDigits,
-      adjustedMinFractionDigits,
+      adjustedMinFractionDigits
     ),
-    minimumFractionDigits: adjustedMinFractionDigits,
     ...options,
   });
 }
@@ -39,7 +38,7 @@ export interface FormatNumberSuffixOptions extends FormatNumberOptions {
 
 export function formatNumberWithSuffix(
   value: number | string,
-  { nullValue, suffix = "", ...options }: FormatNumberSuffixOptions = {},
+  { nullValue, suffix = "", ...options }: FormatNumberSuffixOptions = {}
 ) {
   let parsedNumber = isNumber(value) ? value : Number.parseFloat(value);
   if (Number.isNaN(parsedNumber)) {
@@ -71,7 +70,7 @@ export type NumberLike = null | number | string | undefined;
 export function stringToNumber(value: number | string, nanValue?: NumberLike) {
   if (isNumber(value)) return value;
   const number = Number.parseFloat(
-    isString(value) ? value.replaceAll(",", "") : value,
+    isString(value) ? value.replaceAll(",", "") : value
   );
   if (Number.isNaN(number)) return nanValue;
   return number;